refactor(viewer): modify modifyConnectingLine location to Annotation
