+++ libs/insight-viewer/src/Annotation/utils/modifyConnectingLine.spec.ts
