+++ libs/insight-viewer/src/Annotation/utils/modifyConnectingLine.ts
